ArmVirtKrun: build with strict alignment
When running in EL2 without MMU, all memory is considered to
have device type, which means all accesses must be aligned.

ArmVirtQemu overrides CC_XIPFLAGS to disable the -mstrict-align
compiler flag. This worked fine in QEMU because it either runs
in EL1 with MMU enabled or, when running in EL2, TCG didn't
generate a fault on unaligned accesses.

This doesn't work in neither newer versions on QEMU nor when
running bare metal or with nested virtualization.

In ArmVirtKrun, let's avoid overriding CC_XIPFLAGS to ensure
we build EDK2 with strict alignment.

Signed-off-by: Sergio Lopez <slp@redhat.com>